+++ CHANGELOG.md
@@ -2,6 +2,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Added
+- Added methods to make requests to the data endpoint by [@jquass](https://github.com/jquass)
 ## Changed
 - Removed ally_codes param from get_zetas and changed name to zetas by [@jquass](https://github.com/jquass)
 - Removed ally_codes param from get_squads and changed name to squads by [@jquass](https://github.com/jquass)